Added available asteroid summary and caching Rebase part 1 Rebase 2 part 2
