+++ src/main/java/com/gtnewhorizons/gtnhintergalactic/recipe/IG_Recipe.java
@@ -1,16 +1,9 @@
 package com.gtnewhorizons.gtnhintergalactic.recipe;
 
-import java.util.Arrays;
-import java.util.Map;
-import java.util.Objects;
-import java.util.stream.Collector;
-import java.util.stream.Collectors;
-
 import net.minecraft.item.ItemStack;
 import net.minecraftforge.fluids.FluidStack;
 
 import gregtech.api.util.GTRecipe;
-import gregtech.api.util.GTUtility;
 
 /**
  * GT recipes of GTNH-Intergalactic. These include a needed space project optionally.
@@ -110,7 +103,6 @@ public class IG_Recipe extends GTRecipe {
      */
     public static class IG_SpaceMiningRecipe extends IG_Recipe {
 
-        public String asteroidName;
         public int minDistance;
         public int maxDistance;
         public int minSize;
@@ -118,10 +110,9 @@ public class IG_Recipe extends GTRecipe {
         public int computation;
         public int recipeWeight;
 
-        public IG_SpaceMiningRecipe(boolean aOptimize, String asteroidName, ItemStack[] aItemInputs,
-                ItemStack[] aItemOutputs, FluidStack[] aFluidInputs, int[] aChances, int aDuration, int aEUt,
-                int computation, int minModuleTier, int minDistance, int maxDistance, int minSize, int maxSize,
-                int recipeWeight) {
+        public IG_SpaceMiningRecipe(boolean aOptimize, ItemStack[] aItemInputs, ItemStack[] aItemOutputs,
+                FluidStack[] aFluidInputs, int[] aChances, int aDuration, int aEUt, int computation, int minModuleTier,
+                int minDistance, int maxDistance, int minSize, int maxSize, int recipeWeight) {
             super(
                     aOptimize,
                     aItemInputs,
@@ -133,22 +124,12 @@ public class IG_Recipe extends GTRecipe {
                     aDuration,
                     aEUt,
                     minModuleTier);
-            this.asteroidName = asteroidName;
             this.minDistance = minDistance;
             this.maxDistance = maxDistance;
             this.minSize = minSize;
             this.maxSize = maxSize;
             this.computation = computation;
             this.recipeWeight = recipeWeight;
-        }
-
-        /**
-         * Get the non localized name of the asteroid, can be used in nei and gui
-         * 
-         * @return Asteroid Name
-         */
-        public String getAsteroidName() {
-            return asteroidName;
         }
 
         /**
@@ -159,119 +140,5 @@ public class IG_Recipe extends GTRecipe {
         public int getRecipeWeight() {
             return recipeWeight;
         }
-
-        /**
-         * Compute most of the hash code. The default `IG_SpaceMiningRecipe.hashCode` makes some unchecked assumptions
-         * (asteroidName is unique for output set), so `fullHashCode` exists to bypass these assumptions in case we want
-         * to verify they hold
-         */
-        private int baseHashCode() {
-            int res = 0;
-            res = 31 * res + minDistance;
-            res = 31 * res + maxDistance;
-            res = 31 * res + minSize;
-            res = 31 * res + maxSize;
-            res = 31 * res + computation;
-            res = 31 * res + recipeWeight;
-            res = 31 * res + mSpecialValue;
-            res = 31 * res + mDuration;
-            res = 31 * res + mEUt;
-            return res;
-        }
-
-        /**
-         * Compute the hash code, assuming that asteroidName is the same if and only if output item sets are the same.
-         * Even if this is false, it is still correct for two objects that are not `.equals` to have the same hash code.
-         */
-        public int hashCode() {
-            return 31 * 31 * baseHashCode() + 31 * GTUtility.ItemId.createWithoutNBT(mInputs[0]).hashCode()
-                    + asteroidName.hashCode();
-        }
-
-        /**
-         * Compute the hash code, including an order-invariant hash of the output item set. ONLY USE THIS IF YOU ARE
-         * TESTING RECIPE GENERATION OR SOMETHING. It is unnecessarily expensive most of the time, just use `.hashCode`
-         */
-        public int fullHashCode() {
-            int res = baseHashCode();
-            res = 31 * res + GTUtility.ItemId.createWithoutNBT(mInputs[0]).hashCode();
-            // We don't care about the order of the output items, so we compute the first five sums
-            // of powers of the hashes of the items. This is obviously order invariant, but highly sensitive
-            // to changes of item hashes, which is what we want. Five is more than we need but whatever
-            int[] moments = Arrays.stream(mOutputs).reduce(new int[5], (a, item) -> {
-                int ph = item == null ? 7 : GTUtility.ItemId.createNoCopy(item).hashCode();
-                int x = ph;
-                for (int i = 0; i < a.length; ++i) {
-                    a[i] += x;
-                    x *= ph;
-                }
-                return a;
-            }, (a, b) -> {
-                for (int i = 0; i < a.length; ++i) {
-                    a[i] += b[i];
-                }
-                return a;
-            });
-            for (int ph : moments) {
-                res = 31 * res + ph;
-            }
-            return res;
-        }
-
-        /**
-         * Determine if two recipes are possibly equal. The default `IG_SpaceMiningRecipe.equals` makes some unchecked
-         * assumptions (asteroidName is unique for output set), so `fullEquals` exists to bypass these assumptions in
-         * case we want to verify they hold
-         */
-        private boolean baseEquals(Object _other) {
-            if (!(_other instanceof IG_SpaceMiningRecipe)) {
-                return false;
-            }
-            IG_SpaceMiningRecipe other = (IG_SpaceMiningRecipe) _other;
-            if (minDistance != other.minDistance || maxDistance != other.maxDistance
-                    || minSize != other.minSize
-                    || maxSize != other.maxSize
-                    || computation != other.computation
-                    || recipeWeight != other.recipeWeight
-                    || mSpecialValue != other.mSpecialValue
-                    || mEUt != other.mEUt) {
-                return false;
-            }
-            return true;
-        }
-
-        /**
-         * Check if two space mining recipes are equal, assuming that asteroidName is the same if and only if output
-         * item sets are the same. This should always be the case.
-         */
-        public boolean equals(Object _other) {
-            if (!baseEquals(_other)) {
-                return false;
-            }
-            IG_SpaceMiningRecipe other = (IG_SpaceMiningRecipe) _other;
-            return asteroidName.equals(other.asteroidName)
-                    && GTUtility.ItemId.createWithoutNBT(mInputs[0]).equals(other.mInputs[0]);
-        }
-
-        /**
-         * Check if two space mining recipes are equal, including an order-invariant comparison of the output item set.
-         * ONLY USE THIS IF YOU ARE TESTING RECIPE GENERATION OR SOMETHING. It is unnecessarily expensive most of the
-         * time, just use `.equals`
-         */
-        public boolean fullEquals(Object _other) {
-            if (!baseEquals(_other)) {
-                return false;
-            }
-            IG_SpaceMiningRecipe other = (IG_SpaceMiningRecipe) _other;
-            Collector<ItemStack, ?, Map<GTUtility.ItemId, Long>> collector = Collectors
-                    .toMap(GTUtility.ItemId::createNoCopy, input -> (long) input.stackSize, (a, b) -> a + b);
-            if (!Arrays.stream(mInputs).filter(Objects::nonNull).collect(collector)
-                    .equals(Arrays.stream(other.mInputs).filter(Objects::nonNull).collect(collector))) {
-                return false;
-            }
-            return Arrays.stream(mOutputs).filter(Objects::nonNull).collect(collector)
-                    .equals(Arrays.stream(other.mOutputs).filter(Objects::nonNull).collect(collector));
-        }
     }
 }
-
